requirements.txt
psycopg[binary]==3.3.3
SQLAlchemy==2.0.49
pandas==3.0.2
python-dotenv==1.2.2
jsonschema==4.26.0
PyYAML==6.0.3